otel-python-apps/app-a/requirements.txt
flask
requests

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp

opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
